+++ app/Http/Controllers/ProviderController.php
@@ -11,47 +11,81 @@ class ProviderController extends Controller
 {
     public function index()
     {
-      $providers = Fornecedor::all();
-      return view('providers.list', compact('providers'));
+      if(Auth()->user()->tipo == 'admin' && Auth()->user()->idAdmin != null && Auth()->user()->admin->superAdmin){
+        $providers = Fornecedor::all();
+        return view('providers.list', compact('providers'));
+      }else{
+          /* não tem permissões */
+          abort (401);
+      }
     }
 
     public function create()
     {
-      $provider = new Fornecedor;
-      return view('providers.add', compact('provider'));
+      if(Auth()->user()->tipo == 'admin' && Auth()->user()->idAdmin != null && Auth()->user()->admin->superAdmin){
+        $provider = new Fornecedor;
+        return view('providers.add', compact('provider'));
+      }else{
+          /* não tem permissões */
+          abort (401);
+      }
     }
 
     public function store(StoreProviderRequest $providerRequest)
     {
-      $fields = $providerRequest->validated();
-      $provider = new Fornecedor;
-      $provider->fill($fields);
-      $provider->save();
-      return redirect()->route('provider.index')->with('success', 'Novo fornecedor criado com sucesso.');
+      if(Auth()->user()->tipo == 'admin' && Auth()->user()->idAdmin != null && Auth()->user()->admin->superAdmin){
+        $fields = $providerRequest->validated();
+        $provider = new Fornecedor;
+        $provider->fill($fields);
+        $provider->save();
+        return redirect()->route('provider.index')->with('success', 'Novo fornecedor criado com sucesso.');
+      }else{
+          /* não tem permissões */
+          abort (401);
+      }
     }
 
     public function show(Fornecedor $provider)
     {
-      return view('providers.show', compact('provider'));
+      if(Auth()->user()->tipo == 'admin' && Auth()->user()->idAdmin != null && Auth()->user()->admin->superAdmin){
+        return view('providers.show', compact('provider'));
+      }else{
+          /* não tem permissões */
+          abort (401);
+      }
     }
 
     public function edit(Fornecedor $provider)
     {
-      return view('providers.edit', compact('provider'));
+      if(Auth()->user()->tipo == 'admin' && Auth()->user()->idAdmin != null && Auth()->user()->admin->superAdmin){
+        return view('providers.edit', compact('provider'));
+      }else{
+          /* não tem permissões */
+          abort (401);
+      }
     }
 
     public function update(UpdateProviderRequest $providerRequest, Fornecedor $provider)
     {
-      $fields = $providerRequest->validated();
-      $provider->fill($fields);
-      $provider->updated_at=time();
-      $provider->save();
-      return redirect()->route('provider.index')->with('success', 'Fornecedor editado com sucesso.');
+      if(Auth()->user()->tipo == 'admin' && Auth()->user()->idAdmin != null && Auth()->user()->admin->superAdmin){
+        $fields = $providerRequest->validated();
+        $provider->fill($fields);
+        $provider->save();
+        return redirect()->route('provider.index')->with('success', 'Fornecedor editado com sucesso.');
+      }else{
+          /* não tem permissões */
+          abort (401);
+      }
     }
 
     public function destroy(Fornecedor $provider)
     {
-      $provider->delete();
-      return redirect()->route('provider.index')->with('success', 'Fornecedor eliminado com sucesso');
+      if(Auth()->user()->tipo == 'admin' && Auth()->user()->idAdmin != null && Auth()->user()->admin->superAdmin){
+        $provider->delete();
+        return redirect()->route('provider.index')->with('success', 'Fornecedor eliminado com sucesso');
+      }else{
+          /* não tem permissões */
+          abort (401);
+      }
     }
 }
